Fix camel-casing filenames and remove filename conflicts
