+++ bfxapi/client.py
@@ -7,8 +7,8 @@ a websocket client and a rest interface client
 
 import asyncio
 
-from .websockets.BfxWebsocket import BfxWebsocket
-from .rest.BfxRest import BfxRest
+from .websockets.bfx_websocket import BfxWebsocket
+from .rest.bfx_rest import BfxRest
 
 REST_HOST = 'https://api-pub.bitfinex.com/v2'
 WS_HOST = 'wss://api-pub.bitfinex.com/ws/2'
