+++ bfxapi/utils/decimal.py
