+++ bfxapi/websockets/wallet_manager.py
